+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 ## [1.1.0] - 2024-09-11
 
  - Use url as parameter
+  
+## [2.0.0] - 2024-09-16
+
+ - Change Client class to Reader class
+ - Added option class